requirements.txt
Django>=5
psycopg2-binary
django-redis
